Add files via upload
Customer churn refers to the rate at which customers stop doing business with a company or cancel their subscriptions to its services or products. It is a crucial metric for businesses as it directly affects their revenue and profitability. So, we done a scrutiny on this and provide some solution how we can detact customer churn by using machine learning.